Move app capability handlers to partial class
